\name{as}
\alias{as}
\title{as("classify", "data.frame")}
\description{
  as("classify", "data.frame")

  as("pva", "data.frame")
}
\seealso{
  Other classify: \code{\link{classify.MM}},
  \code{\link{classify.V}}, \code{\link{classify.VA}},
  \code{\link{classify.VI}}

  Other classify: \code{\link{classify.MM}},
  \code{\link{classify.V}}, \code{\link{classify.VA}},
  \code{\link{classify.VI}}
}

