% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bceFun}
\alias{bceFun}
\title{Bivariate Contour Ellipse Function}
\usage{
bceFun(data, level = 0.9, segments = 51)
}
\arguments{
\item{data}{a data frame of x,y coordinates}

\item{level}{the proportion of points to encompass with the ellipse}

\item{segments}{the number of points to divide the ellipse into}
}
\description{
Creates a confidence ellipse around a set of points
}

