\name{chull_area}
\alias{chull_area}
\title{Convex Hull Area}
\usage{
  chull_area(x)
}
\arguments{
  \item{x}{coordinate vectors of points. This can be
  specified as a 2-column matrix x, a list x with two
  components}
}
\value{
  an area (pixels^2)
}
\description{
  Computes the area of minimal convex polygon that spans
  all fixations.
}

