% Generated by roxygen2 (4.0.2): do not edit by hand
\name{circleFun}
\alias{circleFun}
\title{Circle Function}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 50, round = FALSE)
}
\arguments{
\item{center}{the center point of the circle}

\item{diameter}{the diameter of the circle}

\item{npoints}{the number of points to use in the circle}

\item{round}{if TRUE, points will be rounded to nearest integer}
}
\value{
a data frame with columns for the x and y component of each point in the circle
}
\description{
Create a circle from comprised of limited set of points.
}
\examples{
require(ggplot2)
c <- circleFun(center = c(5, 5), diameter = 2)
ggplot(c) + geom_point(aes(x=x,y=y)) + coord_equal()
}

