\name{classify.as.data.frame}
\alias{as.data.frame,classify,missing,missing-method}
\alias{classify.as.data.frame}
\title{Coerce object of class \code{classify} to a Data Frame}
\arguments{
  \item{x}{an object of class
  \code{\linkS4class{classify}}}
}
\description{
  Coerce object of class \code{classify} to a Data Frame
}

