\docType{methods}
\name{getFixations}
\alias{getFixations}
\alias{getFixations,classify,pva-method}
\title{Get Fixations}
\arguments{
  \item{class}{an object of class \code{"classify"}}

  \item{dpva}{an object of class \code{"pva"}}
}
\description{
  Extracts fixation locations from \code{classify} and
  \code{pva} objects.
}

