% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{classify}
\alias{classify}
\title{Classify Gaze}
\usage{
classify(v, e, vt, sigma, minsacdur)
}
\arguments{
\item{v}{a vector of the instantaneous velocities for a set raw gaze samples}

\item{e}{a vector indicating blinks or bad data in the velocity vector}

\item{vt}{saccade onset velocity threshold}

\item{sigma}{when greater than 0, the saccade onset velocity threshold is iteratively adjusted such that
the saccade onset threshold is sigma standard deviations higher than the mean of all velocity samples
lower than the saccade onset threshold, when glissade detection is enabled sigma/2 is used as the
saccade offset velocity threshold}

\item{minsacdur}{the minimum saccade duration in terms of number of samples (to obtain divide desired duration by samplerate)}
}
\description{
Velocity based classification of raw gaze samples to discrete events such as saccades and fixations.
}

