% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{coord_monitor}
\alias{coord_monitor}
\title{Monitor Coordinates}
\usage{
coord_monitor(res_width, res_height)
}
\arguments{
\item{res_width}{the x resolution of a monitor}

\item{res_height}{the y resolution of a monitor}
}
\description{
Sets the ggplot apect ration and limits to the bounds of a computer monitor.
}

