\docType{class}
\name{coverage-class}
\alias{coverage-class}
\title{Class "coverage"}
\description{
  A class to hold gridded ROI coverage information
}
\section{Slots}{
  \describe{ \item{\code{.Data}:}{the proportion of fixated
  ROI cells} \item{\code{coverage}:}{matrix of containing
  fixated ROI cells} }
}

