\name{Coverage}
\alias{Coverage}
\title{Coverage}
\usage{
  Coverage(scanpath)
}
\arguments{
  \item{scanpath}{an object of class
  \core{linkS4class{Scanpath}}}
}
\value{
  an object of class \code{\linkS4class{Coverage}}
}
\description{
  Determines the proportion of fixated ROIs given a
  scanpath
}

