% Generated by roxygen2 (4.0.0): do not edit by hand
\name{detect_blinks.PV}
\alias{detect_blinks.PV}
\title{Detect Blinks (pupil velocity)}
\usage{
detect_blinks.PV(pupil, samplerate, wl = 51, pvt = 1000)
}
\arguments{
  \item{pupil}{a vector of class \code{numeric}; the
  vertical diameter of the pupil}

  \item{samplerate}{the samplerate of the eyetracker}

  \item{wl}{window length of smoothing filter}

  \item{pvt}{pupil diameter velocity threhold}
}
\value{
a vector of class \code{logical}; indicates samples that belong to a blink
}
\description{
Detects blinks based on pupil diameter velocity
}
\examples{
data(smi)
d.b <- with(smi, detect_blinks.PV(smi_dyl, 500))
str(d.b)
}

