% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{detect_blinks.SW}
\alias{detect_blinks.SW}
\title{Detect Blinks (static window)}
\usage{
detect_blinks.SW(pupil, samplerate, pad = 50)
}
\arguments{
\item{pupil}{a vector of class \code{numeric}; the status/validity of the pupil}

\item{samplerate}{the samplerate of the eyetracker}

\item{pad}{time in ms to pad each side of each no pupil segment}
}
\value{
a vector of class \code{logical}; indicates samples that belong to a blink
}
\description{
Detects blinks based on pupil status/valicity
}
\examples{
data(lc120)
d.b <- with(lc120, detect_blinks.SW(status, 120))
str(d.b)
}

