% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dirichlet_tessellation}
\alias{dirichlet_tessellation}
\title{Dirichlet Tessellation}
\usage{
dirichlet_tessellation(x, y, xrange, yrange)
}
\arguments{
\item{x}{Vector of x coordinates of data points}

\item{y}{Vector of y coordinates of data points}

\item{xrange,yrange}{specify a window of observation with these dimensions}
}
\description{
Calculates the dirichlet tessellation of a set of points
}
\examples{
data(smi)
d.pva <- with(smi, pva(smi_sxl, smi_syl,
                       500, 1680, 1050, 473.76, 296.1,
                       smi_ezl, smi_exl, smi_eyl))
d.c <- classify.V(d.pva@v)
d.f <- getFixations(d.c, d.pva)
d.tess <- dirichlet_tessellation(d.f$fixation.x, d.f$fixation.y, c(315,1365), c(0,1050))
str(d.tess)
}

