% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_2_point}
\alias{distance_2_point}
\title{Distance to Point {C++}}
\usage{
distance_2_point(x, y, rx, ry, sw, sh, ez, ex = as.numeric(c(0)),
  ey = as.numeric(c(0)))
}
\arguments{
\item{x}{the horizontal coordinate of a point on a screen (pixels)}

\item{y}{the vertical coordinate of a point on a screen (pixels)}

\item{rx}{the horizontal resolution of the screen (pixels)}

\item{ry}{the vertical resolution of the screen (pixels)}

\item{sw}{the physical screen width (mm)}

\item{sh}{the physical screen height (mm)}

\item{ez}{the perpendicular distance from the viewer to the screen (mm)}

\item{ex}{the horizontal offset of the viewer from screen center (mm)}

\item{ey}{the vertical offset of the viewer from screen center (mm)}
}
\description{
Takes an x and y screen coordinate and returns the physical distance
from the observer to that point on the screen.
}
\examples{
distance_2_point(840, 525, 1680, 1050, 473.76, 296.1, 750)
}

