% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{DynamicROI-class}
\alias{DynamicROI-class}
\title{Class "DynamicROI"}
\description{
Contains information about ROIs that change position or shape over time
}
\section{Slots}{

 \describe{
   \item{\code{.Data}}{a list of polygon coordinates}
   \item{\code{times}}{a data frame where each row contains the begin and end time for each ROI position from the main list}
   \item{\code{ID}}{an object of class \code{"character"}; a text based ID}
   \item{\code{layer}}{an object of class \code{"numeric"}; the layer (z-order / depth) of the ROI}
   \item{\code{parent}}{an object of class \code{"character"}; the id of the parent ROI}
 }
}

