% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DynamicROIs}
\alias{DynamicROIs}
\title{Dynamic ROIs}
\usage{
DynamicROIs(start, end, id, coords, layer = 0, parent = NA_character_)
}
\arguments{
\item{start}{a vector of times indicating the start of a ROIs' existance at a particular location}

\item{end}{a vector of times indicating the end of an ROIs' existance at a particular location}

\item{id}{a vector ROI identifiers}

\item{coords}{a vector of ROI coordinates (a vector of x,y points (also a vector) encoded as a JSON string)}

\item{layer}{a vector of layer identifiers for which each ROI belongs}

\item{parent}{a vector of parent ROI ids which indicates nesting of ROIs}
}
\description{
A helper function for generating a DynamicROIs object.
}

