\name{extent}
\alias{extent}
\title{Extent (range) of fixations}
\usage{
  extent(x)
}
\arguments{
  \item{x}{coordinate vectors of points. This can be
  specified as a 2-column matrix x, a list x with two
  components}
}
\description{
  Calculates the distance between fixations in the
  horizontal and vertical meridians
}

