% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.ROIs}
\alias{fortify.ROIs}
\title{Fortify ROIs}
\usage{
\method{fortify}{ROIs}(rois)
}
\arguments{
\item{rois}{a ROIs object}
}
\value{
a data frame
}
\description{
Method to convert a ROIs object into a data frame useful for plotting.
}

