% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.Scanpath}
\alias{fortify.Scanpath}
\title{Fortify Scanpath}
\usage{
\method{fortify}{Scanpath}(scanpath)
}
\arguments{
\item{scanpath}{a Scanpath object}
}
\value{
a data frame
}
\description{
Method to convert a Scanpath object into a data frame useful for plotting.
}

