% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaze_velocity.R
\name{gaze_velocity}
\alias{gaze_velocity}
\title{Gaze Velocity}
\usage{
gaze_velocity(x, y, ez, ex = 0, ey = 0, timestamp = -1, blinks = NULL,
  samplerate, screen.resolution, screen.dimensions, minsac = 0.02,
  vt = 1000, at = 1e+05)
}
\arguments{
\item{x}{the horizontal coordinate of a point on a screen (pixels)}

\item{y}{the vertical coordinate of a point on a screen (pixels)}

\item{ez}{the perpendicular distance from the viewer to the screen (mm)}

\item{ex}{the horizontal offset of the viewer from screen center (mm)}

\item{ey}{the vertical offset of the viewer from screen center (mm)}

\item{timestamp}{vector of class \code{"numeric"} containing the external time corresponding to raw gaze samples}

\item{blinks}{a vector of class \code{logical} indicating blinks or otherwise bad data in the velocity vector}

\item{samplerate}{the number of samples taken in one second of time}

\item{screen.resolution}{vector of class \code{"numeric"}, the resolution of the screen (pixels)}

\item{screen.dimensions}{vector of class \code{"numeric"}, the physical dimensions of the screen (mm)}

\item{minsac}{the minimum saccade duration in seconds}

\item{vt}{the max peak velocity allowed for good data}

\item{at}{the max peak acceleration allowed for good data}
}
\description{
Computes instantaneous gaze velocity using a Savitzky-Golay filter.
}
\examples{
data(smi)
d <- with(smi, gaze_velocity(smi_sxl, smi_syl,
                             smi_ezl, smi_exl, smi_eyl,
                             timestamp=smi_time, blinks=(smi_dyl==0),
                             500, 1680, 1050, 473.76, 296.1)
d

d <- with(highspeed, gaze_velocity(x,y,1250,1024,768,.38,.30,.67,
                         blinks=(x==0|y==0)))
d
}

