\name{gazetools-package}
\alias{gazetools}
\docType{package}

\title{
  Gaze Tools
}

\description{
  A colletion of tools for processing and classifying eye gaze data.
}

\details{
  \tabular{ll}{
    Package: \tab eyetracking\cr
    Type: \tab Package\cr
    Version: \tab 2.0\cr
    Date: \tab 2012-11-21\cr
    License: \tab GPL-3\cr
  }
}

\author{
  Ryan M. Hope
  Maintainer: Ryan M. Hope <rmh3093@gmail.com>
}

\keyword{eyetracking}
