% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gazetools-class.R
\docType{class}
\name{gazetools}
\alias{gazetools}
\title{Gazetools Reference Class}
\description{
The Gazetools reference class is the main entry point for gaze classification and analysis.
}
\section{Fields}{

\describe{
\item{\code{data}}{an object of class \code{"data.table"} that holds gaze-related timeseries information.}

\item{\code{samplerate}}{the rate at which the gaze data were sampled.}

\item{\code{rx}}{the horizontal resolution of the screen (pixels)}

\item{\code{ry}}{the vertical resolution of the screen (pixels)}

\item{\code{sw}}{the physical screen width (mm)}

\item{\code{sh}}{the physical screen height (mm)}
}}
\section{Methods}{

\describe{
\item{\code{initialize(x, y, samplerate, rx, ry, sw, sh, ez, ex = 0, ey = 0,
  timestamp = -1, order = 2, window = 19, vt = 1000, at = 1e+05,
  blinks = NULL, import = NULL)}}{@parm x this is a param}
}}
\examples{
data(smi)
g <- with(smi, gazetools(smi_sxl,smi_syl, 500,
                         1680, 1050, 473.76, 296.1,
                         smi_ezl, smi_exl, smi_eyl,
                         blinks=(smi_dyl==0|smi_dyr==0)))
}

