\name{geom_bbox}
\alias{geom_bbox}
\title{Bounding Box Geom for ggplot2}
\usage{
  geom_bbox(x, ...)
}
\arguments{
  \item{x}{coordinate vectors of points. This can be
  specified as a 2-column matrix x, a list x with two
  components}

  \item{...}{extra arguments passed on to geom_path}
}
\description{
  Bounding Box Geom for ggplot2
}

