\name{geom_chull}
\alias{geom_chull}
\title{Convex Hull Geom for ggplot2}
\usage{
  geom_chull(x, ...)
}
\arguments{
  \item{x}{coordinate vectors of points. This can be
  specified as a 2-column matrix x, a list x with two
  components}

  \item{...}{extra arguments passed on to geom_path}
}
\description{
  Convex Hull Geom for ggplot2
}

