\name{geom_coverage}
\alias{geom_coverage}
\title{Coverage Geom for ggplot2}
\usage{
  geom_coverage(rois, cv, ...)
}
\arguments{
  \item{rois}{an object of class
  \code{\linkS4class{gridded_rois}}}

  \item{cv}{an object of class
  \code{\linkS4class{coverage}}}

  \item{...}{extra arguments passed on to geom_rect}
}
\description{
  Coverage Geom for ggplot2
}

