\name{geom_grid}
\alias{geom_grid}
\title{Grid Geom for ggplot2}
\usage{
  geom_grid(xintercept, yintercept, ...)
}
\arguments{
  \item{xintercept}{a vector of x intercepts}

  \item{yintercept}{a vector of y intercepts}

  \item{...}{extra arguments passed on to geom_segment}
}
\description{
  Grid Geom for ggplot2
}

