\name{geom_gridded_rois}
\alias{geom_gridded_rois}
\title{Gridded ROIs Geom for ggplot2}
\usage{
  geom_gridded_rois(rois, ...)
}
\arguments{
  \item{rois}{an object of class
  \code{\linkS4class{gridded_rois}}}

  \item{...}{extra arguments passed on to geom_segment}
}
\description{
  Gridded ROIs Geom for ggplot2
}

