\name{geom_roi_coverage}
\alias{geom_roi_coverage}
\title{ROI Coverage Geom for ggplot2}
\usage{
  geom_roi_coverage(rois, cv, ...)
}
\arguments{
  \item{rois}{an object of class
  \code{\linkS4class{gridded_rois}}}

  \item{cv}{an object of class
  \code{\linkS4class{roi_coverage}}}

  \item{...}{extra arguments passed on to geom_rect}
}
\description{
  ROI Coverage Geom for ggplot2
}

