% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{Get Transition Matrix}
\usage{
getTransitionMatrix(scanpath, expected = F, chisquare = F, prob = F,
  digits = 3)
}
\arguments{
\item{scanpath}{a vector a ROI ids (aka scanpath)}

\item{expected}{if TRUE calculate the expected frequencies}

\item{chisquare}{if TRUE calculate the chi-squared statistic}

\item{prob}{if TRUE convert frequencies into probabilities}

\item{digits}{the number of digits to round to}
}
\value{
an object of class \code{\linkS4class{Scanpath}}
}
\description{
Calculates a transition matrix from a scanpath.
}

