\name{geom_png}
\alias{geom_png}
\title{PNG geom}
\usage{
  geom_png(file, xlim = NULL, ylim = NULL)
}
\arguments{
  \item{file}{the path to a png image}

  \item{xlim}{a vector of length 2 that contains the min
  and max horizontal bounds of the image}

  \item{ylim}{a vector of length 2 that contains the min
  and max vertical bounds of the image}
}
\description{
  Creates a ggplot layer that contains a png image
}

