\docType{class}
\name{gridded_rois-class}
\alias{gridded_rois-class}
\title{Class "gridded_rois"}
\description{
  A class to hold gridded ROI information
}
\section{Slots}{
  \describe{ \item{\code{x}:}{vector of class
  \code{"numeric"}, containing the x intercepts for
  vertical grid lines} \item{\code{y}:}{vector of class
  \code{"numeric"}, containing the y intercepts for
  horizontal grid lines} \item{\code{deltax}:}{the width of
  each grid cell} \item{\code{deltay}:}{the height of each
  grid cell} \item{\code{ncol}:}{the number of columns}
  \item{\code{nrow}:}{the number of rows} }
}

