% Generated by roxygen2 (4.0.2): do not edit by hand
\name{local_maxima}
\alias{local_maxima}
\title{Local Maxima}
\usage{
local_maxima(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
a vector of local maxima indices
}
\description{
Finds local maxima in a vector.
}
\examples{
local_maxima(c(1,2,3,2,3,4,3,2,1,2,3,2,1))
}

