\name{mouldThreshold}
\alias{mouldThreshold}

\title{
  Mould Threshold
}

\description{
  A nonparametric method for determining an optimal velocity threshold.
}
\usage{
  mouldThreshold(velocity, samplerate)
}

\arguments{
  \item{velocity}{
    A vector of gaze sample velocities
  }
  \item{samplerate}{
    The samping rate of the gaze samples
  }
}

\value{
  A "mould" object. ~Needs better description.
}

\author{
  Ryan M. Hope
}

\examples{
  # Need an example
}