% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{pva-class}
\alias{pva-class}
\title{Class "pva"}
\description{
A class to hold position, velocity and acceleration of raw gaze data
}
\section{Slots}{

\describe{
\item{\code{.Data}}{object of class \code{"data.table"}, containing the raw gaze data as well as instantaneous velocity and acceleration values}

\item{\code{rx}}{the x resolution of the monitor (pixels)}

\item{\code{ry}}{the y resolution of the monitor (pixels)}

\item{\code{sw}}{the physical screen width (mm)}

\item{\code{sh}}{the physical screen height (mm)}

\item{\code{samplerate}}{the samplerate of the eyetracker}

\item{\code{sgolayfilt}}{parameters for the savitzky-Golay smoothing filter}
}}

