\docType{methods}
\name{plot.pva}
\alias{plot,pva,missing-method}
\alias{plot,voronoi_skewness,missing-method}
\alias{plot.pva}
\alias{plot.voronoi_skewness}
\title{Plot pva}
\arguments{
  \item{x}{an object of class \code{"pva"}}

  \item{y}{an object of class \code{"classify"} (optional)}

  \item{x}{an object of class \code{"voronoi_skewness"}}
}
\description{
  Plot the pva class

  Plot the voronoi_skewness class
}

