\docType{class}
\name{roi_coverage-class}
\alias{roi_coverage-class}
\title{Class "roi_coverage"}
\description{
  A class to hold gridded ROI coverage information
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{the proportion of fixated
  ROI cells} \item{\code{coverage}}{matrix of containing
  fixated ROI cells} }
}

