% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.ROIs}
\alias{plot,ROIs,missing-method}
\alias{plot.ROIs}
\title{Plot ROIs}
\usage{
\S4method{plot}{ROIs,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{ROIs}}
}
\description{
Plots ROIs as filled polygons
}

