\name{scale_grid}
\alias{scale_grid}
\title{Grid Scales for ggplot2}
\usage{
  scale_grid(xintercept, yintercept, outer = T)
}
\arguments{
  \item{xintercept}{a vector of x intercepts}

  \item{yintercept}{a vector of y intercepts}

  \item{outer}{boolean, if TRUE xintercept and yintercept
  are assumed to contain intercepts for the plot limits}
}
\description{
  Discrete grid scales
}

