\docType{class}
\name{Scanpath-class}
\alias{Scanpath-class}
\title{Class "Scanpath"}
\description{
Holds ROIs and a vector of ROI IDs in the order they were fixated.
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{a vector of class
  \code{"character"}; fixated ROI IDs}
  \item{\code{ROIs}}{an object of class \code{"ROIs"}; the
  set of ROIs} }
}

