% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigthresh}
\alias{sigthresh}
\title{Sigma Threshold {C++}}
\usage{
sigthresh(x, e, threshold, sigma)
}
\arguments{
\item{x}{a vector of values that a sigma threshold needs to be determined for}

\item{e}{a vector of class \code{"logical"} indicating samples that should be excluded from any calculations}

\item{threshold}{the starting threshold for the iterative algorithm}

\item{sigma}{the number of standard deviations of separation between the threshold and the mean of the values below the threhold}
}
\description{
Iteratively converges on a threshold for which the mean of the values below the threshold
is \emph{sigma} standard deviations away from the threshold
}
\examples{
d <- with(highspeed, pva(x,y,1250,1024,768,.38,.30,.67))
sigthresh(d$v,d$blinks,80,5)
}

