% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{skewness.tess}
\alias{skewness.tess}
\title{Tessellation Skewness}
\usage{
skewness.tess(x)
}
\arguments{
\item{x}{an object of class \code{"tess"}}
}
\value{
the skewness of the tesselation tile areas
}
\description{
Calculates the skewness of the tesselation tile areas
}

