% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uidvec}
\alias{uidvec}
\title{Logical-based (non-continuous) unique ID runs {C++}}
\usage{
uidvec(x)
}
\arguments{
\item{x}{a vector of class \code{"logical"}}
}
\value{
a vector of \code{"numeric"}
}
\description{
For a vector of logicals, assigns each continuous run of TRUE values a unique ID.
}
\examples{
uidvec(c(FALSE,FALSE,TRUE,TRUE,TRUE,FALSE,TRUE,FALSE,TRUE,TRUE,FALSE,TRUE))
}

