\docType{class}
\name{voronoi_skewness-class}
\alias{voronoi_skewness-class}
\title{Class "voronoi_skewness"}
\description{
  A class to hold Voronoi skewness information
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{the skewness of the area
  of the voronoi cells} \item{\code{polygons}}{an object of
  class \code{\linkS4class{SpatialPolygonsDataFrame}}} }
}

