\docType{class}
\name{VoronoiPolygons-class}
\alias{VoronoiPolygons-class}
\title{Class "VoronoiPolygons"}
\description{
A class to hold Voronoi tessellation information
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{an object of class
  \code{\linkS4class{ROIs}}} \item{\code{skewness}}{the
  skewness of the area of the voronoi cells} }
}

