% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.VoronoiPolygons}
\alias{plot,VoronoiPolygons,missing-method}
\alias{plot.VoronoiPolygons}
\title{Plot VoronoiPolygons}
\usage{
\S4method{plot}{VoronoiPolygons,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{VoronoiPolygons}}}
}
\description{
Plots Voronoi Tessellation
}
\examples{
data(smi)
d.pva <- with(smi, pva(smi_sxl, smi_syl,
                       500, 1680, 1050, 473.76, 296.1,
                       smi_ezl, smi_exl, smi_eyl))
d.c <- classify.V(d.pva@v)
d.f <- getFixations(d.c, d.pva)
d.vp <- VoronoiPolygons(d.f[,c("fixation.x","fixation.y")], c(315,1365,0,1050))
plot(d.vp)
}

