% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VoronoiPolygons}
\alias{VoronoiPolygons}
\title{VoronoiPolygons}
\usage{
VoronoiPolygons(x, rw)
}
\arguments{
\item{x}{a matrix of x,y cooridinates}

\item{rw}{the coordinates of the corners of the rectangular window enclosing the voronoi cells, in the order (xmin, xmax, ymin, ymax).}
}
\value{
an object of class \code{\linkS4class{VoronoiPolygons}}
}
\description{
Creates an object of class VoronoiPolygons
}
\examples{
data(smi)
d.pva <- with(smi, pva(smi_sxl, smi_syl,
                       500, 1680, 1050, 473.76, 296.1,
                       smi_ezl, smi_exl, smi_eyl))
d.c <- classify.V(d.pva@v)
d.f <- getFixations(d.c, d.pva)
d.vp <- VoronoiPolygons(d.f[,c("fixation.x","fixation.y")], c(315,1365,0,1050))
str(d.vp)
}

