% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as}
\alias{as}
\title{as("classify", "data.frame")}
\description{
as("classify", "data.frame")

Force a 'pva' Object to Belong to a Class
}
\seealso{
Other classify: \code{\link{classify.MM}};
  \code{\link{classify.VA}}; \code{\link{classify.VI}};
  \code{\link{classify.V}}
}

