% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{classify-class}
\alias{classify-class}
\title{Class "classify"}
\description{
A class to hold gaze data classifications
}
\section{Slots}{

\describe{
\item{\code{.Data}}{object of class \code{"data.table"}, containing the gaze sample classification information}

\item{\code{algorithm}}{the algorithm used to classify the gaze data}

\item{\code{thresholds}}{the threshold settings for the classification algorithm}
}}

