% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.classify}
\alias{plot,classify,pva-method}
\alias{plot.classify}
\title{Plot classify}
\usage{
\S4method{plot}{classify,pva}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{classify}}

\item{y}{an object of class \code{pva}}
}
\description{
Creates a scatter plot of the scanpath between each fixation. Fixation durations are indicated by the size of the fixation point.
}
\examples{
data(smi)
d.pva <- with(smi, pva(smi_sxl, smi_syl,
                       500, 1680, 1050, 473.76, 296.1,
                       smi_ezl, smi_exl, smi_eyl, pupil=smi_dxl))
d.c <- classify.VA(d.pva@v, d.pva@a, blinks=d.pva@blinks)
plot(d.c, d.pva)
}

