\docType{methods}
\name{classify.as.data.frame}
\alias{as.data.frame,classify,missing,missing-method}
\alias{classify.as.data.frame}
\alias{getFixations}
\alias{getFixations,classify,pva-method}
\alias{plot,classify,pva-method}
\alias{plot.classify}
\title{Plot classify}
\arguments{
  \item{x}{an object of class \code{"classify"}}

  \item{y}{an object of class \code{"pva"}}

  \item{class}{an object of class \code{"classify"}}

  \item{dpva}{an object of class \code{"pva"}}
}
\description{
  Plot the classify class

  Extracts fixation locations from \code{classify} and
  \code{pva} objects.
}

