\docType{class}
\name{Coverage-class}
\alias{Coverage-class}
\title{Class "Coverage"}
\description{
  Holds a Scanpath and the proportion of ROIs fixated
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{an object of class
  \code{"numeric"}; the proportion of ROIs fixated}
  \item{\code{Scanpath}}{an object of class
  \code{"Scanpath"}; the scanpath} }
}

