\name{coverage}
\alias{coverage}
\title{Coverage}
\usage{
  coverage(rois, fixations)
}
\arguments{
  \item{rois}{an object of class \code{gridded_rois}}

  \item{fixations}{coordinate vectors of fixations. This
  can be specified as a 2-column matrix x, a list x with
  two components}
}
\description{
  Calculates the proportion of ROI cells covered by
  fixations
}

