% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gazetools-class.R
\docType{class}
\name{gazetools-class}
\alias{gazetools}
\alias{gazetools-class}
\title{Gazetools Reference Class}
\description{
The Gazetools reference class is the main entry point for gaze classification and analysis.
}
\section{Fields}{

\describe{
\item{\code{data}}{an object of class \code{"data.table"} that holds gaze-related timeseries information.}

\item{\code{samplerate}}{the rate at which the gaze data were sampled.}

\item{\code{rx}}{the horizontal resolution of the screen (pixels)}

\item{\code{ry}}{the vertical resolution of the screen (pixels)}

\item{\code{sw}}{the physical screen width (mm)}

\item{\code{sh}}{the physical screen height (mm)}
}}
\section{Methods}{

\describe{
\item{\code{initialize(..., import = NULL)}}{Initialize object with raw gaze data. Takes same arguments as \code{\link{pva}}.}
}}
\examples{
data(smi)
g <- with(smi, gazetools(smi_sxl,smi_syl, 500,
                         1680, 1050, 473.76, 296.1,
                         smi_ezl, smi_exl, smi_eyl,
                         blinks=(smi_dyl==0|smi_dyr==0),
                         timestamp=smi_time))
g

g <- with(highspeed, gazetools(x,y,1250,1024,768,.38,.30,.67,
                               blinks=(x==0|y==0)))
g
}

