% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazetools-package.R
\docType{package}
\name{gazetools-package}
\alias{gazetools-package}
\title{Gazetools}
\description{
A collection of tools for classifying, analyzing and visualizing eye gaze data.
}
\author{
Ryan M. Hope \email{rmh3093@gmail.com}
}

