% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getScanpath}
\alias{getScanpath}
\title{Get Scanpath}
\usage{
getScanpath(x, y = NULL, begin = NULL, end = NULL, rois,
  nearest = FALSE, firstOnly = TRUE)
}
\arguments{
\item{x,y}{the x and y coordinates of a set of points. Alternatively, a single argument x can be provided}

\item{begin,end}{the begin and end timestamp of each fixation, only used with DynamicROIs}

\item{rois}{an object of class \code{\linkS4class{ROIs}} or \code{\linkS4class{DynamicROIs}}}

\item{nearest}{if TRUE fixations not on an ROI will be moved to the nearest ROI}

\item{firstOnly}{if TRUE only DynamicROIs present at the start of a fixation are reported}
}
\value{
an object of class \code{\linkS4class{Scanpath}}
}
\description{
Determines the scanpath given a set of fixations and a list of polygon ROIs
}

