\docType{class}
\name{mould-class}
\alias{mould-class}
\title{Class "mould"}
\description{
A class to hold mould threshold information
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{a threshold value}
  \item{\code{thresholds}}{vector of class
  \code{"numeric"}} \item{\code{resp1}}{vector of class
  \code{"numeric"}} \item{\code{resp2}}{vector of class
  \code{"numeric"}} \item{\code{gap}}{vector of class
  \code{"numeric"}} }
}

