% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mouldThreshold}
\alias{mouldThreshold}
\title{Mould Threshold}
\usage{
mouldThreshold(v, samplerate)
}
\arguments{
  \item{v}{a vector of instantaneous velocity values}

  \item{samplerate}{eyetracker samplerate}
}
\value{
an object of class \code{mould}
}
\description{
Uses the Mould algorithm to determine the optimal velocity or acceleration threshold for
classifying raw gaze data.
}
\examples{
data(smi)
d.pva <- with(smi, pva(smi_sxl, smi_syl,
                       500, 1680, 1050, 473.76, 296.1,
                      smi_ezl, smi_exl, smi_eyl))
d.mt <- mouldThreshold(d.pva@v, d.pva@samplerate)
str(d.mt)
}
\references{
Mould, M. S., Foster, D. H., Amano, K., & Oakley, J. P. (2012). A simple nonparametric method for classifying eye fixations. Vision research, 57, 18-25.
}

