% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{pva}
\alias{pva}
\title{Position, Velocity, Acceleration}
\usage{
pva(x, y, samplerate, rx, ry, sw, sh, ez, ex = 0, ey = 0, timestamp = -1,
  minsac = 0.02, vt = 1000, at = 1e+05, blinks = NULL)
}
\arguments{
\item{x}{the horizontal coordinate of a point on a screen (pixels)}

\item{y}{the vertical coordinate of a point on a screen (pixels)}

\item{samplerate}{the number of samples taken in one second of time}

\item{rx}{the horizontal resolution of the screen (pixels)}

\item{ry}{the vertical resolution of the screen (pixels)}

\item{sw}{the physical screen width (mm)}

\item{sh}{the physical screen height (mm)}

\item{ez}{the perpendicular distance from the viewer to the screen (mm)}

\item{ex}{the horizontal offset of the viewer from screen center (mm)}

\item{ey}{the vertical offset of the viewer from screen center (mm)}

\item{timestamp}{vector of class \code{"numeric"} containing the external time corresponding to raw gaze samples}

\item{minsac}{the minimum saccade duration in seconds}

\item{vt}{the max peak velocity allowed for good data}

\item{at}{the max peak acceleration allowed for good data}

\item{blinks}{a vector of class \code{logical} indicating blinks or otherwise bad data in the velocity vector}
}
\description{
Takes standard eye tracker gaze information and computes smoothed position, velocity
and acceleration profiles using a Savitzky-Golay filter.
}
\examples{
data(smi)
d <- with(smi, pva(smi_sxl,smi_syl, 500,
                   1680, 1050, 473.76, 296.1,
                   smi_ezl, smi_exl, smi_eyl,
                   blinks=(smi_dyl==0|smi_dyr==0),
                   timestamp=smi_time))
d

d <- with(highspeed, pva(x,y,1250,1024,768,.38,.30,.67,
                         blinks=(x==0|y==0)))
d

}

