\name{pvaPlot}
\alias{pvaPlot}

\title{
  Position, Velocity, Acceleration Plot
}

\description{
  Takes a pva data.frame and plots the x and y gaze position as well as the 
  instantaneous velocity and acceleration over time.
}
\usage{
  pvaPlot(d)
}

\arguments{
  \item{d}{
    A pva data.frame returned by pva()
  }
}

\author{
  Ryan M. Hope
}

\examples{
  # Need an example
}