\name{ROI}
\alias{ROI}
\title{Region of Interest (ROI)}
\usage{
ROI(coords, ID)
}
\arguments{
  \item{coords}{a data frame of xy coordinates}

  \item{ID}{a text based identifier for the new ROI}
}
\value{
an object of class \code{\linkS4class{ROI}}
}
\description{
Creates a polygon based region of interest from a set of coordinates.
}

