% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rplot}
\alias{rplot}
\title{Raster Plot}
\usage{
rplot(raster, xlim = NULL, ylim = NULL, rois = NULL, ...)
}
\arguments{
\item{raster}{a rasterized image}

\item{xlim}{a vector of length 2 that contains the min and max horizontal bounds of the image}

\item{ylim}{a vector of length 2 that contains the min and max vertical bounds of the image}

\item{rois}{an object of class \code{"ROIs"}, used to overlay ROI polygons}

\item{...}{extra params passed to geom_path, only used when rois!=NULL}
}
\description{
A wrapper around a few ggplot2 functions that creates a base plot layer out of a raster image.
Coordinates are restricted to values within the raster image and the aspect ratio is fixed.
}
\examples{
data(logo)
rplot(as.raster(logo))
}

