% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ruidvec}
\alias{ruidvec}
\title{String-based (continuous) unique ID runs {C++}}
\usage{
ruidvec(x)
}
\arguments{
\item{x}{a vector of class \code{"character"}}
}
\value{
a vector of \code{"numeric"}
}
\description{
For a vector strings, assigns each continuous run of particular string a unique ID.
}
\examples{
ruidvec(c("One","One","Two","Three","Three","Two","Three","Three","One"))
}

