\name{voronoi_polygons}
\alias{voronoi_polygons}
\title{Voronoi Polygons}
\usage{
  voronoi_polygons(x, rw)
}
\arguments{
  \item{x}{a matrix of x,y cooridinates}

  \item{rw}{the coordinates of the corners of the
  rectangular window enclosing the voronoi cells, in the
  order (xmin, xmax, ymin, ymax).}
}
\description{
  Computes Voronoi polygons based on a set of x,y
  coordinates
}

